pyproject.toml
[project]
name = "yet-another-pipeline-test-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-bigquery>=3.37.0",
    "google-cloud-storage>=3.3.1",
    "requests>=2.32.5",
]
